ml/requirements.txt
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.9
boto3==1.34.0
joblib==1.3.2
structlog==24.1.0
prometheus-client==0.19.0
shap==0.44.0
optuna==3.5.0
imbalanced-learn==0.11.0
